Initial commit of stand-alone lnic generator repo.
Copied the chisel source files over from the lnic/rocket-chip repo.

Need to move some of the classes back into the rocket-chip repo, such
as the TX and RX queues, which go in the RocketCore CSR file.

Also need to update the Top level (and tile level) mixins along with
the configs to align with what the latest chipyard repo expects.